requirements.txt
Flask
fpdf
python-docx
python-dotenv
requests
docx2txt
PyMuPDF
